Refactor to internal package
